若干の手直し
